# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/pools": {
            "get": {
                "description": "List pools in pagination view.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pool"
                ],
                "summary": "List pools",
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "The number of skipped records, usually it's pageSize * (pageNumber - 1)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "The number of records displayed on the page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort in ASC or DESC order by sortField",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "tvl",
                            "trade",
                            "liquidity"
                        ],
                        "type": "string",
                        "default": "tvl",
                        "description": "The field used for sorting. The value is tvl, trade or liquidity",
                        "name": "sortField",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paged pools",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.PagingResult-model_PoolInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "600": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "description": "List users in pagination view.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "List users",
                "parameters": [
                    {
                        "minimum": 0,
                        "type": "integer",
                        "default": 0,
                        "description": "The number of skipped records, usually it's pageSize * (pageNumber - 1)",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "The number of records displayed on the page",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort in ASC or DESC order by sortField",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "trade",
                            "liquidity"
                        ],
                        "type": "string",
                        "default": "trade",
                        "description": "The field used for sorting. The value is trade or liquidity",
                        "name": "sortField",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paged users",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.PagingResultWithUpdatedAt-model_UserInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "600": {
                        "description": "Internal server error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.BusinessError"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.BusinessError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code error code, 0 indicates success, otherwise business error (e.g. user not found) or internal server error.",
                    "type": "integer"
                },
                "data": {
                    "description": "Data is the return value if success. Otherwise, it is concret failure reason in string type."
                },
                "message": {
                    "description": "Message error message associated with `Code`.",
                    "type": "string"
                }
            }
        },
        "model.PagingResult-model_PoolInfo": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PoolInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.PagingResultWithUpdatedAt-model_UserInfo": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.UserInfo"
                    }
                },
                "total": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "integer"
                }
            }
        },
        "model.PoolInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "fee": {
                    "type": "integer"
                },
                "liquidityWeight": {
                    "type": "integer"
                },
                "token0": {
                    "type": "string"
                },
                "token0Symbol": {
                    "type": "string"
                },
                "token1": {
                    "type": "string"
                },
                "token1Symbol": {
                    "type": "string"
                },
                "tradeWeight": {
                    "type": "integer"
                },
                "tvl": {
                    "type": "number"
                }
            }
        },
        "model.UserInfo": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "liquidityPoints": {
                    "type": "number"
                },
                "tradePoints": {
                    "type": "number"
                }
            }
        }
    }
}